Explore debug logging
Found earliest dprintf that worked and added some printf in platform.cc.

Determined that the code in `debuglog_dumper` of
`//zircon/kernel/lib/debuglog/debuglog.cc` is
where the timestamp, pid and tid are printed.

First I searched the code for `dprintf` define (figuring it was a macro):

  $ find . -type f -name '*.h' -print | xargs grep -ni '#define dprintf'
  ./third_party/libxml2/trio.h:214:#define dprintf trio_dprintf
  ./zircon/kernel/include/debug.h:69:#define dprintf(level, x...) \

Yep, which leads to printf which leads to _printf which leads to
_printf_engine. Adding outputting the `fmt` in `_printf_engine` of
`//zircon/kernel/lib/libc/printf.c` yields:

  Booting from ROM..[%05d.%03d] %05lu:%05lu> %s
  [00000.000] 00000:00000> zbi: @ %p (%u bytes)
  [%05d.%03d] %05lu:%05lu> %s
  [00000.000] 00000:00000> zbi: @ 0xffffff8000c14000 (11602616 bytes)

So format contains `[%05d.%03d]` and then searching for it yields:

  $ find . -type f -name '*.c' -print -o -type f -name '*.cc' -print -o -type f -name '*.cpp' -print -o -type f -name '*.h' -print | xargs grep -ni '\[%05d\.%03d\]'
  ./src/developer/feedback/feedback_agent/attachments/kernel_log_ptr.cc:85:      kernel_log += fxl::StringPrintf("[%05d.%03d] %05" PRIu64 ".%05" PRIu64 "> %s\n",
  ./src/developer/feedback/feedback_agent/attachments/system_log_ptr.cc:150:  logs_ += fxl::StringPrintf("[%05d.%03d][%05" PRIu64 "][%05" PRIu64 "][%s] %s: %s\n",
  ./zircon/kernel/lib/debuglog/debuglog.cc:379:      n = snprintf(tmp, sizeof(tmp), "[%05d.%03d] %05" PRIu64 ":%05" PRIu64 "> %s\n",
  ./zircon/system/uapp/dlog/dlog.c:98:      size_t len = snprintf(tmp, sizeof(tmp), "[%05d.%03d] ", (int)(rec->timestamp / 1000000000ULL),
  ./zircon/system/core/netsvc/debuglog.cc:56:      snprintf(out, MAX_LOG_LINE, "[%05d.%03d] %05" PRIu64 ".%05" PRIu64 "> %s\n",

Adding `sav1` and `sav2` to the two most probable files:

  ./zircon/kernel/lib/debuglog/debuglog.cc:379:      n = snprintf(tmp, sizeof(tmp), "sav2: [%05d.%03d] %05" PRIu64 ":%05" PRIu64 "> %s\n",
  ./zircon/system/core/netsvc/debuglog.cc:56:      snprintf(out, MAX_LOG_LINE, "sav1: [%05d.%03d] %05" PRIu64 ".%05" PRIu64 "> %s\n",

and found sav2 being printed in the output:

  Booting from ROM..sav2: [%05d.%03d] %05lu:%05lu> %s
  sav2: [00000.000] 00000:00000> zbi: @ %p (%u bytes)
  sav2: [%05d.%03d] %05lu:%05lu> %s
  sav2: [00000.000] 00000:00000> zbi: @ 0xffffff8000c14000 (11602744 bytes)

So `./zircon/kernel/lib/debuglog/debuglog.cc:379` is where the timestamp,
pid and tid are printed!!

I also wanted to find out what how the flow went in main.cc added more
dprintf and changed some debug levels from SPEW to INFO.

Change-Id: Icaae5b9bc2540fb77a6b23a8e3b3fc089242ddd2